Made changes in Showdetails

+++ PCIapi/Model/ExeMaturity.cs
@@ -12,6 +12,6 @@
         public string AreasDesc { get; set; }
         public string ActivityDesc { get; set; }
         public string CompValue { get; set; }
-        public int ScoreValue { get; set; }
+        public decimal ScoreValue { get; set; }
     }
 }
